+++ src/ngMock/angular-mocks.js
@@ -756,6 +756,36 @@ angular.mock.TzDate = function (offset, timestamp) {
 angular.mock.TzDate.prototype = Date.prototype;
 /* jshint +W101 */
 
+// TODO(matias): remove this IMMEDIATELY once we can properly detect the
+// presence of a registered module
+var animateLoaded;
+try {
+  angular.module('ngAnimate');
+  animateLoaded = true;
+} catch(e) {}
+
+if(animateLoaded) {
+  angular.module('ngAnimate').config(['$provide', function($provide) {
+    var reflowQueue = [];
+    $provide.value('$$animateReflow', function(fn) {
+      reflowQueue.push(fn);
+      return angular.noop;
+    });
+    $provide.decorator('$animate', function($delegate) {
+      $delegate.triggerReflow = function() {
+        if(reflowQueue.length === 0) {
+          throw new Error('No animation reflows present');
+        }
+        angular.forEach(reflowQueue, function(fn) {
+          fn();
+        });
+        reflowQueue = [];
+      };
+      return $delegate;
+    });
+  }]);
+}
+
 angular.mock.animate = angular.module('mock.animate', ['ng'])
 
   .config(['$provide', function($provide) {
@@ -1911,7 +1941,6 @@ angular.mock.clearDataCache = function() {
     }
   }
 };
-
 
 
 if(window.jasmine || window.mocha) {